lol I'm sorry, I need to stop trying to take shortcuts
