+++ components/weak-aura-editor.tsx
@@ -4,14 +4,24 @@ import { RubyVM } from "@ruby/wasm-wasi";
 import { useCallback, useContext, useEffect } from "react";
 import { GlobalContext } from "../app/providers";
 import { Button } from "./ui/button";
+import { decodeHash, encodeHash } from "@/lib/utils";
 
 type WeakAuraEditorProps = {
-  ruby: RubyVM;
+  ruby?: RubyVM;
   onChange?: (result: string) => void;
 };
 
+export const defaultSource = `title 'Shadow Priest WhackAura'
+load spec: :shadow_priest
+hide_ooc!
+
+dynamic_group 'WhackAuras' do
+  debuff_missing 'Shadow Word: Pain'
+end`;
+
 export function WeakAuraEditor({ ruby, onChange }: WeakAuraEditorProps) {
   const { source, setSource } = useContext(GlobalContext);
+
   useEffect(() => {
     if (!ruby) return;
     ruby?.evalAsync(rubyInit);
@@ -23,6 +33,25 @@ export function WeakAuraEditor({ ruby, onChange }: WeakAuraEditorProps) {
       onChange?.(result.toString());
     });
   }, [source, ruby]);
+
+  useEffect(() => {
+    if (!source || source === "" || source === defaultSource) return;
+    const base64 = encodeHash(source);
+    if (window?.location) window.location.hash = base64;
+    history.pushState(null, "", `#${base64}`);
+  }, [source]);
+
+  useEffect(() => {
+    const handleHashChange = () => {
+      if (!window?.location?.hash) return;
+      setSource(decodeHash(window?.location?.hash));
+    };
+    window.addEventListener("hashchange", handleHashChange);
+    if (window?.location?.hash) handleHashChange();
+    return () => {
+      window.removeEventListener("hashchange", handleHashChange);
+    };
+  }, []);
 
   return (
     <div className="grid gap-4 w-full">
